requirements.txt
# Web 框架
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9

# 数据库
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.20.0

# RAG / LLM
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
openai>=1.50.0
rank-bm25>=0.2.2
jieba>=0.42.1

# 工具
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
httpx>=0.27.0
